polls save to state when saved
