+++ src/action/room.js
@@ -38,12 +38,3 @@ export const createPollAction = poll => ({
   type: 'POLL_CREATE',
   payload: poll,
 });
-
-export const saveRoomAction = ({ roomData, token }) => (store) =>
-  superagent.post(`${API_URL}/session`)
-    .set('Authorization', `Bearer ${token}`)
-    .send(roomData)
-    .then((data) => {
-      console.log(data);
-    })
-    .catch(console.log);
